Feature account

+++ App/views/single.php
@@ -9,8 +9,10 @@
 </div>
 <br>
 <div class="actions">
-    <a href="../public/index.php?route=editArticle&articleId=<?= $article->getId(); ?>">Modifier</a>
-    <a href="../public/index.php?route=deleteArticle&articleId=<?= $article->getId(); ?>">Supprimer</a>
+    <?php if($this->session->get('role') === 'admin') { ?>    
+        <a href="../public/index.php?route=editArticle&articleId=<?= $article->getId(); ?>">Modifier</a>
+        <a href="../public/index.php?route=deleteArticle&articleId=<?= $article->getId(); ?>">Supprimer</a>
+    <?php } ?>
 </div>
 <br>
 <a href="../public/index.php">Retour à l'accueil</a>
@@ -26,7 +28,7 @@
         <p><?= htmlspecialchars($comment->getContent());?></p>
         <p>Posté le <?= htmlspecialchars($comment->getCreatedAt());?></p>
         <?php
-        var_dump($comment);
+
         if($comment->isFlag()) {
             ?>
             <p>Ce commentaire a déjà été signalé</p>
